change navbar to side bar for menu
