+++ frontend/src/components/Layout.tsx
@@ -1,9 +1,9 @@
-import Navbar from './Navbar';
+import Sidebar from './Sidebar';
 import React from 'react';
 export default function Layout({ children }: { children: React.JSX.Element }) {
   return (
     <>
-      <Navbar />
+      <Sidebar />
       <main className="p-4">{children}</main>
     </>
   );
